refactor: rename `repository` => `storage`
